Move timestamp.js, preprocessor.js to lib/utilities.
