+++ src/app/lib/utilities/preprocessor.js
@@ -1,4 +1,4 @@
-import nodeLogger from '../helpers/logger.node';
+import nodeLogger from '../../helpers/logger.node';
 import applyTimestampRules from './rules/timestamp';
 
 const defaultRules = [applyTimestampRules];
